Refactoring: extract input values from auth provider
